Refactor TransactionHistoryRestController to delegate transaction history and export logic to TransactionHistoryService
